move internal competition header file
