+++ src/competition.hpp
